Make the repository cache live for 1hr instead of forever
I wouldn't suggest for going "forever". I know it sounds cool not having to query the database anymore but, I'd go with a sensible 60min. That's way beyond enough for our use case. This will prevent any error on cached content or if anyone tweaked something directly in the database (which they shouldn't)... it will cover your back when someone complains as if it happened it will simply update the content after 1hr.